requirements.txt
Django==3.2.24
python-decouple==3.5
djangorestframework==3.12.4
djangorestframework-simplejwt
django-countries
psycopg2-binary==2.9.9
requests==2.28.1
cryptography